# lib/api-zod/src/generated/types/appPasswordInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do Foco OS e endpoints protegidos para automações.
 * OpenAPI spec version: 0.1.0
 */

export interface AppPasswordInput {
  /**
     * @minLength 1
     * @maxLength 1024
     */
  password: string;
}