# lib/api-zod/src/generated/types/authenticatedSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do Foco OS e endpoints protegidos para automações.
 * OpenAPI spec version: 0.1.0
 */

export interface AuthenticatedSession {
  authenticated: boolean;
}